+++ t_12/Core/Src/tasks.c
@@ -2,15 +2,15 @@
 #include "interlayer.h"
 
 // Analog part
-extern volatile double m_data_AI_1; // analog input 1
-extern volatile double m_data_AI_2;
-extern volatile double m_data_AI_3;
-extern volatile double m_data_AI_4;
+extern volatile float m_data_AI_1; // analog input 1
+extern volatile float m_data_AI_2;
+extern volatile float m_data_AI_3;
+extern volatile float m_data_AI_4;
 
-// extern volatile double m_data_AO_1; // analog output 1
-// extern volatile double m_data_AO_2;
-// extern volatile double m_data_AO_3;
-// extern volatile double m_data_AO_4;
+// extern volatile float m_data_AO_1; // analog output 1
+// extern volatile float m_data_AO_2;
+// extern volatile float m_data_AO_3;
+// extern volatile float m_data_AO_4;
 
 // Discret part
 extern volatile bool m_data_DI_1; // discret input 1
@@ -26,10 +26,10 @@ extern volatile bool m_data_DO_4;
 
 
 // Step points
-extern volatile double m_data_SP_AI_1; // analog input 1
-extern volatile double m_data_SP_AI_2;
-extern volatile double m_data_SP_AI_3;
-extern volatile double m_data_SP_AI_4;
+extern volatile float m_data_SP_AI_1; // analog input 1
+extern volatile float m_data_SP_AI_2;
+extern volatile float m_data_SP_AI_3;
+extern volatile float m_data_SP_AI_4;
 
 // Types AI --> 50М, 100P, Pt100, Pt1000, L, K, J, S, B, 0..20mA, 4..20mA, 0..10V
 extern volatile uint8_t m_data_T_AI_1;
